Updated install script

+++ install.sh
@@ -4,11 +4,19 @@ if [ "$EUID" -ne 0 ]
  exit
 else
   SCRIPT_DIR=$(dirname $(realpath "$0"))
-  # Creates the binary
+  # Create executable
   cargo build --release
   BINARY_DIR="$SCRIPT_DIR/target/release/baup"
-  # Creates the symbolic link
-  sudo ln -s $BINARY_DIR /usr/bin/baup2
-  ls -l /usr/bin/baup2
+  echo "[0] Create symbolic link"
+  echo "[1] Copy the executable"
+  read -p "Select Option: " SEL
+  if [ $SEL -eq 0 ]; then
+    sudo ln -s $BINARY_DIR /usr/bin/baup2
+    ls -l /usr/bin/baup2
+  elif [ $SEL -eq 1 ]; then
+    copy $BINARY_DIR /usr/bin/baup2
+  else; then
+    echo "Invalid option. Please run the script again"
+  fi
 fi
